SkBee - move addon stuff to new class AddonLoader
